requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
appdirs==1.4.3            # via black
astroid==2.2.5            # via pylint
atomicwrites==1.3.0       # via pytest
attrs==19.1.0             # via black, packaging, pytest
black==19.10b0
bleach==3.1.3
certifi==2019.6.16        # via requests
chardet==3.0.4            # via requests
click==7.0
dict-deep==2.0.2
docutils==0.15.2          # via readme-renderer
htmlmin==0.1.12           # via mkdocs-minify-plugin
idna==2.8                 # via requests
importlib-metadata==0.19  # via pluggy, pytest
invoke==1.3.0
isort==4.3.21
jinja2==2.10.1            # via mkdocs, punch.py
jsmin==2.2.2              # via mkdocs-minify-plugin
kv==0.3
lazy-object-proxy==1.4.2  # via astroid
livereload==2.6.1         # via mkdocs
loguru==0.3.2
markdown==3.1.1           # via mkdocs, pymdown-extensions
markupsafe==1.1.1         # via jinja2
mccabe==0.6.1             # via pylint
melddict==1.0.1
mkdocs-material==4.4.2
mkdocs-minify-plugin==0.2.1  # via mkdocs-material
mkdocs==1.0.4
more-itertools==7.2.0     # via pytest, zipp
packaging==19.1           # via pytest, pytest-sugar
pathspec==0.7.0           # via black
pkginfo==1.5.0.1          # via twine
pluggy==0.12.0            # via pytest
punch.py==1.6.2
py==1.8.0                 # via pytest
pygments==2.4.2
pylint==2.3.1
pymdown-extensions==6.0
pyparsing==2.4.2          # via packaging
pytest-mock==1.10.4
pytest-sugar==0.9.2
pytest==5.0.1
python-dotenv==0.10.3
python-frontmatter==0.4.5
pyyaml-include==1.1.1.1
pyyaml==5.1.2
readme-renderer==24.0     # via twine
regex==2020.2.20          # via black
requests-toolbelt==0.9.1  # via twine
requests==2.22.0          # via requests-toolbelt, twine
sh==1.12.14
six==1.12.0               # via astroid, bleach, livereload, packaging, punch.py, python-frontmatter, readme-renderer
termcolor==1.1.0          # via pytest-sugar
toml==0.10.0              # via black
tornado==6.0.3            # via livereload, mkdocs
tqdm==4.35.0              # via twine
twine==1.13.0
typed-ast==1.4.0          # via astroid, black
urllib3==1.25.3           # via requests
wcwidth==0.1.7            # via pytest
webencodings==0.5.1       # via bleach
wrapt==1.11.2             # via astroid
zipp==0.6.0               # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
